The one with the new fitness

+++ Assets/Scripts/GA/Genome.cs
@@ -9,12 +9,12 @@ public abstract class Genome : IComparable
     public long Lengthother;
     public int CrossoverPoint;
     public int MutationIndex;
-    public float CurrentFitness = 0.0f;
+    public double CurrentFitness = 0.0;
     abstract public void Initialize();
     abstract public void Mutate( );
     abstract public Genome Crossover(Genome g);
     abstract public char GenerateGeneValue( );
-    abstract public float CalculateFitness();
+    abstract public double CalculateFitness();
     abstract public bool CanReproduce(float fitness);
     abstract public void SetCrossoverPoint(int crossoverPoint);
     abstract public bool CanDie(float fitness);
